+++ src/dbn_store.cpp
@@ -1,4 +1,4 @@
-#include "databento/dbn_file_store.hpp"
+#include "databento/dbn_store.hpp"
 
 #include <memory>   // unique_ptr
 #include <utility>  // move
@@ -6,19 +6,22 @@
 #include "databento/file_stream.hpp"
 #include "databento/record.hpp"
 
-using databento::DbnFileStore;
+using databento::DbnStore;
 
-DbnFileStore::DbnFileStore(const std::filesystem::path& file_path)
+DbnStore::DbnStore(const std::filesystem::path& file_path)
     : decoder_{ILogReceiver::Default(), InFileStream{file_path}} {}
 
-DbnFileStore::DbnFileStore(ILogReceiver* log_receiver,
-                           const std::filesystem::path& file_path,
-                           VersionUpgradePolicy upgrade_policy)
+DbnStore::DbnStore(ILogReceiver* log_receiver, const std::filesystem::path& file_path,
+                   VersionUpgradePolicy upgrade_policy)
     : decoder_{log_receiver, std::make_unique<InFileStream>(file_path),
                upgrade_policy} {}
 
-void DbnFileStore::Replay(const MetadataCallback& metadata_callback,
-                          const RecordCallback& record_callback) {
+DbnStore::DbnStore(ILogReceiver* log_receiver, std::unique_ptr<IReadable> input,
+                   VersionUpgradePolicy upgrade_policy)
+    : decoder_{log_receiver, std::move(input), upgrade_policy} {}
+
+void DbnStore::Replay(const MetadataCallback& metadata_callback,
+                      const RecordCallback& record_callback) {
   auto metadata = decoder_.DecodeMetadata();
   if (metadata_callback) {
     metadata_callback(std::move(metadata));
@@ -31,21 +34,21 @@ void DbnFileStore::Replay(const MetadataCallback& metadata_callback,
   }
 }
 
-void DbnFileStore::Replay(const RecordCallback& record_callback) {
+void DbnStore::Replay(const RecordCallback& record_callback) {
   Replay({}, record_callback);
 }
 
-const databento::Metadata& DbnFileStore::GetMetadata() {
+const databento::Metadata& DbnStore::GetMetadata() {
   MaybeDecodeMetadata();
   return metadata_;
 }
 
-const databento::Record* DbnFileStore::NextRecord() {
+const databento::Record* DbnStore::NextRecord() {
   MaybeDecodeMetadata();
   return decoder_.DecodeRecord();
 }
 
-void DbnFileStore::MaybeDecodeMetadata() {
+void DbnStore::MaybeDecodeMetadata() {
   if (!has_decoded_metadata_) {
     metadata_ = decoder_.DecodeMetadata();
     has_decoded_metadata_ = true;